Run Python in the browser via Pyodide (backlog #11)
The spike's verdict is positive, and the feature is built rather than
just recommended.

Why this fits where the AI Tutor didn't: Pyodide is CPython compiled
to WebAssembly and runs entirely in the user's tab. No server, no API
key, no recurring cost — the §5.3 static-first rule is preserved
intact. The only cost is a one-time ~10 MB runtime download, which is
why loading is lazy and never happens until someone presses Run.

Feasibility was measured, not assumed: all 126 Hub samples were
executed in Pyodide. 110 run, median 7 ms, p90 1.8 s. The 16 that
cannot are recorded in src/data/python/runnability.js with a reason
each — PyTorch/XGBoost/LightGBM/CatBoost/arch/Prophet/Gymnasium have
no WebAssembly build (11), two download datasets at runtime, two are
large simulations that exceed a sensible browser timeout, and one
exhausts the tab's memory on a kernel grid search.

That table drives the UI: a Run button appears only where running
genuinely works. Everywhere else the block states the actual reason
and suggests copying the code to run locally — an honest "not here,
because X" beats a button that always fails, and beats hiding the
distinction entirely.

The audit also found a real bug in the shipped content: grp_sd's
np.repeat call passed a pandas int64 array as the repeat count, which
raises a cast error on any 32-bit target — WebAssembly here, but
Windows equally. Fixed with .astype(int) and re-verified in Pyodide;
it now runs and agrees with the hand-computed result.

Implementation notes:
- One shared load promise, so double-clicking Run cannot start two
  10 MB downloads; a failed load clears it so a retry is possible.
- stdout is redirected per run and restored in a finally-equivalent
  path, so one failing sample cannot silently swallow later output.
- Execution errors are returned, not thrown: a failing sample is a
  normal outcome for an output panel to display, not an exception
  that should trip the error boundary added in 163da5e.
- matplotlib is switched to the Agg backend at load, since a sample
  calling savefig() would otherwise raise on a headless canvas.

10 tests cover the table and error formatting, including two that
fail loudly if a future sample imports an unavailable package or
downloads a dataset without being marked — the ways this feature
would silently start lying.

Verified in-browser: Run on Pearson r downloaded the runtime and
printed the real computed output (r = 0.929, p = 6.08e-18), a second
run reused the cached runtime, and an LSTM sample correctly shows
"PyTorch is not available in the browser runtime" with no Run button.

162 tests pass.

+++ src/data/python/pyStatsDescriptive.js
@@ -605,8 +605,11 @@ s2 = (n * sum_fx2 - sum_fx ** 2) / (n * (n - 1))
 print(f"grouped mean = {sum_fx/n:.2f} min")
 print(f"grouped s    = {np.sqrt(s2):.2f} min  (n = {n})")
 
-# Equivalent modern one-liner: expand midpoints by frequency
-expanded = np.repeat(tbl["Xm"].values, tbl["f"].values)
+# Equivalent modern one-liner: expand midpoints by frequency.
+# .astype(int) keeps this portable — np.repeat requires a platform-native
+# integer count, and a pandas int64 column raises a cast error on 32-bit
+# targets (Windows, and WebAssembly runtimes like Pyodide).
+expanded = np.repeat(tbl["Xm"].values, tbl["f"].values.astype(int))
 print(f"check via repeat: s = {expanded.std(ddof=1):.2f}")`,
     scenario: {
       title: 'Commute-time spread from a survey table',
